Add empty placeholder within starred collections
