+++ shared/i18n/locales/en_US/translation.json
@@ -138,11 +138,12 @@
   "Documents": "Documents",
   "Logo": "Logo",
   "Document archived": "Document archived",
+  "Empty": "Empty",
+  "Move document": "Move document",
   "Collections": "Collections",
   "You can't reorder documents in an alphabetically sorted collection": "You can't reorder documents in an alphabetically sorted collection",
   "Untitled": "Untitled",
   "New nested document": "New nested document",
-  "Move document": "Move document",
   "Document not supported – try Markdown, Plain text, HTML, or Word": "Document not supported – try Markdown, Plain text, HTML, or Word",
   "Starred documents could not be loaded": "Starred documents could not be loaded",
   "Starred": "Starred",
